// frontcars/src/types.ts
import { SVGProps } from "react";

export interface ICar {
  auction: string;
  bid: number;
  color: string;
  condition: string;
  docs: string;
  drive: string;
  engine: string;
  fuel: string;
  imagesrc: any;
  location: string;
  mileage: number;
  model: string;
  note: any;
  primarydam: string;
  saledate: string;
  secondarydam: string;
  seller: string;
  trans: string;
  vin: string;
  year: number;
}